SECURITY: Correctly set Content-disposition header
- If Content-disposition header is set to inline, then malicious files stored in
the secured folder can be run in the logged-in user's context (e.g. HTML files
that run some AJAX to modify things in the CMS). This change forces the
'attachment' disposition by default that ensures the browser is forced to
download these files rather than displaying them inline. A further fix might be
to use mime validation instead of a single-purpose rule, to allow e.g. image
files to use inline. New configuration option added to allow the default to be
changed if required.

Thanks to Chris McCurley of Aura Information Security for reporting this issue.